% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-tables.R
\name{add_criteria_headers}
\alias{add_criteria_headers}
\title{Function to add criteria headers (removes criteria and time columns)}
\usage{
add_criteria_headers(hux_table, data, n_cols, time_digits = 1)
}
\arguments{
\item{hux_table}{A hux table to modify}

\item{data}{the original data.frame version of the above table}

\item{n_cols}{Number of columns}

\item{time_digits}{Number of decimal places for expected analysis time}
}
\value{
Modified hux table
}
\description{
Function to add criteria headers (removes criteria and time columns)
}
