% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-tables.R
\name{add_hypothesis_headers}
\alias{add_hypothesis_headers}
\title{Function to add hypothesis header rows (removes hypothesis column)}
\usage{
add_hypothesis_headers(hux_table, data, n_cols)
}
\arguments{
\item{hux_table}{A hux table to modify}

\item{data}{the original data.frame version of the above table}

\item{n_cols}{Number of columns}
}
\value{
Modified hux table
}
\description{
Function to add hypothesis header rows (removes hypothesis column)
}
