% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-sim.R
\name{apply_hypothesis_tests}
\alias{apply_hypothesis_tests}
\title{Apply hypothesis tests to analyses}
\usage{
apply_hypothesis_tests(analyses, hypotheses)
}
\arguments{
\item{analyses}{Data frame containing analysis results}

\item{hypotheses}{Data frame containing hypothesis definitions}
}
\description{
Apply hypothesis tests to analyses
}
