% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-sim.R
\name{calculate_analysis_timing_cpp_wrapper}
\alias{calculate_analysis_timing_cpp_wrapper}
\title{Wrapper function to fix C++ list-column issue}
\usage{
calculate_analysis_timing_cpp_wrapper(df, analyses, hypotheses)
}
\arguments{
\item{df}{Data frame with simulation data}

\item{analyses}{Data frame containing analysis definitions}

\item{hypotheses}{Data frame containing hypothesis definitions}
}
\description{
Wrapper function to fix C++ list-column issue
}
