% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-sim.R
\name{calculate_z_bounds}
\alias{calculate_z_bounds}
\title{Calculate z bounds}
\usage{
calculate_z_bounds(
  info_levels = c(100, 200, 300, 400, 500)/4,
  alpha = 0.025,
  sfu = gsDesign::sfLDOF,
  sfupar = NULL,
  planned_max_info = 500/4
)
}
\arguments{
\item{info_levels}{Information levels for analyses}

\item{alpha}{Type I error rate}

\item{sfu}{Spending function}

\item{sfupar}{Spending function parameters}

\item{planned_max_info}{Planned maximum information}
}
\description{
Calculate z bounds
}
