% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{check_allowed_columns}
\alias{check_allowed_columns}
\title{Check allowed columns}
\usage{
check_allowed_columns(df, name, allowed_cols)
}
\arguments{
\item{df}{Data.frame to check}

\item{name}{Name of data.frame for error messages}

\item{allowed_cols}{Vector of allowed column names}
}
\description{
Check allowed columns
}
