% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{check_analyses_dependencies}
\alias{check_analyses_dependencies}
\title{Check analyses dependencies}
\usage{
check_analyses_dependencies(config)
}
\arguments{
\item{config}{Full config for cross-references}
}
\description{
Check analyses dependencies
}
