% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{check_analyses_structure}
\alias{check_analyses_structure}
\title{Check analyses data.frame structure}
\usage{
check_analyses_structure(analyses)
}
\arguments{
\item{analyses}{Analyses data.frame}
}
\description{
Check analyses data.frame structure
}
