% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{check_bounded_double_column}
\alias{check_bounded_double_column}
\title{Check bounded double column}
\usage{
check_bounded_double_column(
  df,
  col_name,
  lower_bound,
  upper_bound,
  strict = FALSE
)
}
\arguments{
\item{df}{Data.frame}

\item{col_name}{Column name}

\item{lower_bound}{Lower bound}

\item{upper_bound}{Upper bound}

\item{strict}{If TRUE, bounds are exclusive}
}
\description{
Check bounded double column
}
