% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{check_character_vector_or_na_column}
\alias{check_character_vector_or_na_column}
\title{Check character vector or NA column (list column)}
\usage{
check_character_vector_or_na_column(df, col_name)
}
\arguments{
\item{df}{Data.frame}

\item{col_name}{Column name}
}
\description{
Check character vector or NA column (list column)
}
