% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{check_column_all_or_none}
\alias{check_column_all_or_none}
\title{Check that column is present in all objects or none}
\usage{
check_column_all_or_none(objects, col_name)
}
\arguments{
\item{objects}{Named list of data.frames}

\item{col_name}{Column name to check}
}
\description{
Check that column is present in all objects or none
}
