% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{check_column_all_or_none_mapped}
\alias{check_column_all_or_none_mapped}
\title{Check that mapped columns are present in all objects or none}
\usage{
check_column_all_or_none_mapped(objects, col_map, col_desc)
}
\arguments{
\item{objects}{Named list of data.frames}

\item{col_map}{Named list mapping object names to column names}

\item{col_desc}{Description for error message}
}
\description{
Check that mapped columns are present in all objects or none
}
