% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{check_column_presence}
\alias{check_column_presence}
\title{Check that all objects have specified column}
\usage{
check_column_presence(objects, col_name, context)
}
\arguments{
\item{objects}{Named list of data.frames}

\item{col_name}{Column name to check}

\item{context}{Context for error message}
}
\description{
Check that all objects have specified column
}
