% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{check_config_by_element}
\alias{check_config_by_element}
\title{Check if config is a valid list with required structure}
\usage{
check_config_by_element(config)
}
\arguments{
\item{config}{Configuration object to validate}
}
\value{
TRUE if valid, throws error otherwise
}
\description{
Check if config is a valid list with required structure
}
