% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{check_config_column_consistency}
\alias{check_config_column_consistency}
\title{Check column consistency across config components}
\usage{
check_config_column_consistency(config)
}
\arguments{
\item{config}{Configuration list}
}
\description{
Check column consistency across config components
}
