% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{check_config_dependencies}
\alias{check_config_dependencies}
\title{Check dependencies between config components}
\usage{
check_config_dependencies(config)
}
\arguments{
\item{config}{Configuration list}
}
\description{
Check dependencies between config components
}
