% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{check_distribution_bin_structure}
\alias{check_distribution_bin_structure}
\title{Check distribution_bin data.frame structure}
\usage{
check_distribution_bin_structure(distribution_bin)
}
\arguments{
\item{distribution_bin}{Distribution_bin data.frame}
}
\description{
Check distribution_bin data.frame structure
}
