% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{check_distribution_combinations}
\alias{check_distribution_combinations}
\title{Check distribution combinations exist}
\usage{
check_distribution_combinations(
  endpoint,
  strata,
  treatments,
  dist_tte,
  dist_bin,
  field_name,
  row_num
)
}
\arguments{
\item{endpoint}{Endpoint name}

\item{strata}{Strata vector}

\item{treatments}{Treatments vector}

\item{dist_tte}{TTE distribution data.frame}

\item{dist_bin}{Binary distribution data.frame}

\item{field_name}{Field name for error messages}

\item{row_num}{Row number for error messages}
}
\description{
Check distribution combinations exist
}
