% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{check_distribution_presence}
\alias{check_distribution_presence}
\title{Check at least one distribution is present}
\usage{
check_distribution_presence(config)
}
\arguments{
\item{config}{Configuration list}
}
\description{
Check at least one distribution is present
}
