% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{check_distribution_tte_structure}
\alias{check_distribution_tte_structure}
\title{Check distribution_tte data.frame structure}
\usage{
check_distribution_tte_structure(distribution_tte)
}
\arguments{
\item{distribution_tte}{Distribution_tte data.frame}
}
\description{
Check distribution_tte data.frame structure
}
