% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{check_double_or_null_column}
\alias{check_double_or_null_column}
\title{Check double or NULL column}
\usage{
check_double_or_null_column(df, col_name)
}
\arguments{
\item{df}{Data.frame}

\item{col_name}{Column name}
}
\description{
Check double or NULL column
}
