% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{check_enroll_rate_dependencies}
\alias{check_enroll_rate_dependencies}
\title{Check enroll_rate dependencies}
\usage{
check_enroll_rate_dependencies(config)
}
\arguments{
\item{config}{Full config}
}
\description{
Check enroll_rate dependencies
}
