% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{check_enroll_rate_structure}
\alias{check_enroll_rate_structure}
\title{Check enroll_rate data.frame structure}
\usage{
check_enroll_rate_structure(enroll_rate)
}
\arguments{
\item{enroll_rate}{Enroll_rate data.frame}
}
\description{
Check enroll_rate data.frame structure
}
