% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{check_graph_dependencies}
\alias{check_graph_dependencies}
\title{Check graph dependencies}
\usage{
check_graph_dependencies(config)
}
\arguments{
\item{config}{Full config}
}
\description{
Check graph dependencies
}
