% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{check_graph_structure}
\alias{check_graph_structure}
\title{Check graph list structure}
\usage{
check_graph_structure(graph)
}
\arguments{
\item{graph}{Graph list}
}
\description{
Check graph list structure
}
