% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{check_hypotheses_dependencies}
\alias{check_hypotheses_dependencies}
\title{Check hypotheses dependencies}
\usage{
check_hypotheses_dependencies(config)
}
\arguments{
\item{config}{Full config for cross-references}
}
\description{
Check hypotheses dependencies
}
