% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{check_hypotheses_structure}
\alias{check_hypotheses_structure}
\title{Check hypotheses data.frame structure}
\usage{
check_hypotheses_structure(hypotheses)
}
\arguments{
\item{hypotheses}{Hypotheses data.frame}
}
\description{
Check hypotheses data.frame structure
}
