% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{check_nonnegative_double_column}
\alias{check_nonnegative_double_column}
\title{Check non-negative double column}
\usage{
check_nonnegative_double_column(df, col_name)
}
\arguments{
\item{df}{Data.frame}

\item{col_name}{Column name}
}
\description{
Check non-negative double column
}
