% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{check_optional_character}
\alias{check_optional_character}
\title{Check optional character field}
\usage{
check_optional_character(config, field_name)
}
\arguments{
\item{config}{Configuration list}

\item{field_name}{Name of field to check}
}
\description{
Check optional character field
}
