% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{check_optional_date}
\alias{check_optional_date}
\title{Check optional date field}
\usage{
check_optional_date(config, field_name)
}
\arguments{
\item{config}{Configuration list}

\item{field_name}{Name of field to check}
}
\description{
Check optional date field
}
