% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{check_positive_double_or_inf_column}
\alias{check_positive_double_or_inf_column}
\title{Check positive double or Inf column}
\usage{
check_positive_double_or_inf_column(df, col_name)
}
\arguments{
\item{df}{Data.frame}

\item{col_name}{Column name}
}
\description{
Check positive double or Inf column
}
