% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{check_positive_double_vector_or_null_column}
\alias{check_positive_double_vector_or_null_column}
\title{Check positive double vector or NULL column (list column)}
\usage{
check_positive_double_vector_or_null_column(df, col_name)
}
\arguments{
\item{df}{Data.frame}

\item{col_name}{Column name}
}
\description{
Check positive double vector or NULL column (list column)
}
