% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{check_positive_integer_or_na_column}
\alias{check_positive_integer_or_na_column}
\title{Check positive integer or NA column}
\usage{
check_positive_integer_or_na_column(df, col_name)
}
\arguments{
\item{df}{Data.frame}

\item{col_name}{Column name}
}
\description{
Check positive integer or NA column
}
