% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{check_required_dataframe}
\alias{check_required_dataframe}
\title{Check required data.frame field}
\usage{
check_required_dataframe(config, field_name)
}
\arguments{
\item{config}{Configuration list}

\item{field_name}{Name of field to check}
}
\description{
Check required data.frame field
}
