% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{check_required_list}
\alias{check_required_list}
\title{Check required list field}
\usage{
check_required_list(config, field_name)
}
\arguments{
\item{config}{Configuration list}

\item{field_name}{Name of field to check}
}
\description{
Check required list field
}
