% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{check_sf_sfpar}
\alias{check_sf_sfpar}
\title{Check sf and sfpar combination}
\usage{
check_sf_sfpar(sf, sfpar, n_analyses, alpha, row_num)
}
\arguments{
\item{sf}{Spending function}

\item{sfpar}{Spending function parameter}

\item{n_analyses}{Number of analyses}

\item{alpha}{Alpha level}

\item{row_num}{Row number for error messages}
}
\description{
Check sf and sfpar combination
}
