% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_MT_in_grSeq.R
\name{checkInput}
\alias{checkInput}
\title{Title}
\usage{
checkInput(inputD, G, enrollment = NULL)
}
\arguments{
\item{inputD}{An input dataset}

\item{G}{A graph object}

\item{enrollment}{An enrollment object}
}
\value{
A list of objects for output tables and plots
}
\description{
Title
}
\examples{
\dontrun{
checkInput(inputD,G)
}
}
