% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_study.R
\name{create_study}
\alias{create_study}
\title{Create new study from config and template}
\usage{
create_study(config, rmd_template = "gsd_detailed", output_dir = ".")
}
\arguments{
\item{config}{Either a config list object, config name from repository, or path to .R config file}

\item{rmd_template}{Either a template name or path to .Rmd template file (default: "gsd_detailed")}

\item{output_dir}{Directory to create study files}
}
\description{
Create new study from config and template
}
