% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process.R
\name{create_summary_tables}
\alias{create_summary_tables}
\title{Create summary tables}
\usage{
create_summary_tables(
  analyses,
  hypotheses,
  graph,
  weights,
  sim_results = NULL,
  alpha = 0.025
)
}
\arguments{
\item{analyses}{Processed analyses}

\item{hypotheses}{Processed hypotheses}

\item{graph}{Graph object}

\item{weights}{Weight scenarios}

\item{sim_results}{Simulation results (optional)}

\item{alpha}{Overall Type I error rate}
}
\value{
List of summary tables
}
\description{
Create summary tables
}
