% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_MT_in_grSeq.R
\name{exec_calc}
\alias{exec_calc}
\title{Processing util function that generate a content given input}
\usage{
exec_calc(inputD)
}
\arguments{
\item{inputD}{An input dataset with specs.}
}
\value{
A list with main objects to be outputed: D, ia_details and hyp_testing_dataset
}
\description{
Processing util function that generate a content given input
}
\examples{
\dontrun{
  # Need to define inputD first
  exec_calc(inputD)
}
}
