% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-process.R
\name{expected_d_at_t}
\alias{expected_d_at_t}
\title{Expected events at specific time}
\usage{
expected_d_at_t(t = 10, fail_rate, enroll_rate)
}
\arguments{
\item{t}{Time point of interest}

\item{fail_rate}{Failure rate information}

\item{enroll_rate}{Enrollment information}
}
\value{
Expected events
}
\description{
Expected events at specific time
}
