% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-process.R
\name{expected_n_at_t}
\alias{expected_n_at_t}
\title{Expected sample size at specific time}
\usage{
expected_n_at_t(
  t = 10,
  maturity_time = 6,
  enroll_rate = gsDesign2::define_enroll_rate(duration = 10, rate = 10)
)
}
\arguments{
\item{t}{Time point of interest}

\item{maturity_time}{Endpoint maturity time}

\item{enroll_rate}{Enrollment information}
}
\value{
Expected sample size
}
\description{
Expected sample size at specific time
}
