% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-process.R
\name{expected_t_at_d}
\alias{expected_t_at_d}
\title{Expected time for specific number of events}
\usage{
expected_t_at_d(d = 20, fail_rate, enroll_rate, interval = c(0.001, 1e+06))
}
\arguments{
\item{d}{Events of interest}

\item{fail_rate}{Failure rate information}

\item{enroll_rate}{Enrollment information}

\item{interval}{Interval for root finding}
}
\value{
Expected time
}
\description{
Expected time for specific number of events
}
