% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-process.R
\name{expected_t_at_n}
\alias{expected_t_at_n}
\title{Expected time for specific sample size}
\usage{
expected_t_at_n(
  n = 100,
  maturity_time = 6,
  enroll_rate = gsDesign2::define_enroll_rate(duration = 10, rate = 10)
)
}
\arguments{
\item{n}{Sample size of interest}

\item{maturity_time}{Endpoint maturity time}

\item{enroll_rate}{Enrollment information}
}
\value{
Expected time
}
\description{
Expected time for specific sample size
}
