% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-sim.R
\name{extract_information_levels}
\alias{extract_information_levels}
\title{Extract information levels from analyses results}
\usage{
extract_information_levels(analyses_result, hypotheses)
}
\arguments{
\item{analyses_result}{Results from hypothesis testing}

\item{hypotheses}{Data frame containing hypothesis definitions}
}
\description{
Extract information levels from analyses results
}
