% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reporting.R
\name{generate_report}
\alias{generate_report}
\title{Generate GSD Report}
\usage{
generate_report(
  gsd_details,
  output_file = NULL,
  template_type = "html",
  template_file = NULL,
  output_dir = NULL
)
}
\arguments{
\item{gsd_details}{Results object from process_config function}

\item{output_file}{Output file path (defaults to config name + extension)}

\item{template_type}{Type of report template ("html", "pdf", "word")}

\item{template_file}{Path to Rmd template file (defaults to package template)}

\item{output_dir}{Output directory (defaults to config file directory)}
}
\description{
Generate GSD Report
}
