% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/templates.R
\name{get_config_path}
\alias{get_config_path}
\title{Get path to config file from repository}
\usage{
get_config_path(config_name)
}
\arguments{
\item{config_name}{Name of config in repository}
}
\value{
Path to config file
}
\description{
Get path to config file from repository
}
