% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-process.R
\name{get_cum_alpha_spend}
\alias{get_cum_alpha_spend}
\title{Extract nominal cum alpha-spending from OC object}
\usage{
get_cum_alpha_spend(oc)
}
\arguments{
\item{oc}{Operating characteristics object}
}
\value{
Vector of cumulative alpha spent
}
\description{
Extract nominal cum alpha-spending from OC object
}
