% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-wlr.R
\name{get_delta_tite_wlr}
\alias{get_delta_tite_wlr}
\title{Compute WLR drift (delta) at a vector of analysis times}
\usage{
get_delta_tite_wlr(
  control,
  test,
  enroll_rate,
  fail_rate,
  times_analysed,
  weight
)
}
\arguments{
\item{control}{Character name of the control treatment.}

\item{test}{Character name of the test treatment.}

\item{enroll_rate}{Processed enrollment data frame.}

\item{fail_rate}{Distribution data frame.}

\item{times_analysed}{Numeric vector of analysis times.}

\item{weight}{Weight argument as returned by
\code{test_method_to_wlr_weight()}.}
}
\value{
Numeric vector of drift values.
}
\description{
Compute WLR drift (delta) at a vector of analysis times
}
\keyword{internal}
