% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-process.R
\name{get_description_effect_size}
\alias{get_description_effect_size}
\title{Get description of effect size}
\usage{
get_description_effect_size(oc)
}
\arguments{
\item{oc}{Operating characteristics object}
}
\value{
Character string describing effect size
}
\description{
Get description of effect size
}
