% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-process.R
\name{get_description_max_info}
\alias{get_description_max_info}
\title{Get description of maximal information level}
\usage{
get_description_max_info(dist_type, max_information_factor, digits = 0)
}
\arguments{
\item{dist_type}{Distribution type}

\item{max_information_factor}{Maximum information factor}

\item{digits}{Number of digits for rounding}
}
\value{
Character string describing max information
}
\description{
Get description of maximal information level
}
