% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-process.R
\name{get_description_sf}
\alias{get_description_sf}
\title{Get description of spending function}
\usage{
get_description_sf(sf, sfpar, nominal)
}
\arguments{
\item{sf}{Spending function}

\item{sfpar}{Spending function parameter}

\item{nominal}{Nominal spend information}
}
\value{
Character string describing spending function
}
\description{
Get description of spending function
}
