% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-process.R
\name{get_description_trigger_long}
\alias{get_description_trigger_long}
\title{Get long description of analysis trigger}
\usage{
get_description_trigger_long(
  endpoint,
  strata,
  treatments,
  sample_size,
  events,
  description_short,
  distribution
)
}
\arguments{
\item{endpoint}{Endpoint name}

\item{strata}{Strata}

\item{treatments}{Treatments}

\item{sample_size}{Sample size}

\item{events}{Number of events}

\item{description_short}{Short description}

\item{distribution}{Distribution object}
}
\value{
Long description string
}
\description{
Get long description of analysis trigger
}
