% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-process.R
\name{get_description_trigger_short}
\alias{get_description_trigger_short}
\title{Get short description of analysis trigger}
\usage{
get_description_trigger_short(endpoint, sample_size, events)
}
\arguments{
\item{endpoint}{Endpoint name}

\item{sample_size}{Sample size}

\item{events}{Number of events}
}
\value{
Short description string
}
\description{
Get short description of analysis trigger
}
