% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{get_enroll_rate_combinations}
\alias{get_enroll_rate_combinations}
\title{Get enroll_rate combinations (unnested)}
\usage{
get_enroll_rate_combinations(enroll_rate)
}
\arguments{
\item{enroll_rate}{Enroll_rate data.frame}
}
\value{
Character vector of combinations
}
\description{
Get enroll_rate combinations (unnested)
}
