% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-process.R
\name{get_hurdles}
\alias{get_hurdles}
\title{Extract exit hurdles from OC object}
\usage{
get_hurdles(oc)
}
\arguments{
\item{oc}{Operating characteristics object}
}
\value{
Vector of exit hurdles
}
\description{
Extract exit hurdles from OC object
}
