% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{get_hypothesis_combinations}
\alias{get_hypothesis_combinations}
\title{Get all hypothesis combinations}
\usage{
get_hypothesis_combinations(hypotheses)
}
\arguments{
\item{hypotheses}{Hypotheses data.frame}
}
\value{
Character vector of combinations
}
\description{
Get all hypothesis combinations
}
