% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-sim.R
\name{get_maurer_bretz}
\alias{get_maurer_bretz}
\title{Implement M/B - version for IPD simulations}
\usage{
get_maurer_bretz(analyses, hypotheses, graph)
}
\arguments{
\item{analyses}{Data frame containing analysis results}

\item{hypotheses}{Data frame containing hypothesis definitions}

\item{graph}{Graph object with edges and weights}
}
\description{
Implement M/B - version for IPD simulations
}
