% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-process.R
\name{get_nominal_p}
\alias{get_nominal_p}
\title{Extract nominal p-value thresholds from OC object}
\usage{
get_nominal_p(oc)
}
\arguments{
\item{oc}{Operating characteristics object}
}
\value{
Vector of nominal p-values
}
\description{
Extract nominal p-value thresholds from OC object
}
