% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-sim.R
\name{get_p_thresholds}
\alias{get_p_thresholds}
\title{Get p thresholds}
\usage{
get_p_thresholds(sf, sfpar, info, alpha, obs_info)
}
\arguments{
\item{sf}{Spending function name}

\item{sfpar}{Spending function parameters}

\item{info}{Information levels}

\item{alpha}{Type I error rate}

\item{obs_info}{Observed information levels}
}
\description{
Get p thresholds
}
