% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-process.R
\name{get_pooled_rate}
\alias{get_pooled_rate}
\title{Calculate pooled rate across strata}
\usage{
get_pooled_rate(enroll_rate, distribution)
}
\arguments{
\item{enroll_rate}{Enrollment rate data}

\item{distribution}{Distribution data}
}
\value{
Pooled rate
}
\description{
Calculate pooled rate across strata
}
