% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-process.R
\name{get_powers}
\alias{get_powers}
\title{Extract cumulative powers from OC object}
\usage{
get_powers(oc)
}
\arguments{
\item{oc}{Operating characteristics object}
}
\value{
Vector of cumulative powers
}
\description{
Extract cumulative powers from OC object
}
