% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/templates.R
\name{get_rmd_template_path}
\alias{get_rmd_template_path}
\title{Get path to Rmd template skeleton}
\usage{
get_rmd_template_path(rmd_template_name)
}
\arguments{
\item{rmd_template_name}{Name of Rmd template}
}
\value{
Path to skeleton.Rmd file
}
\description{
Get path to Rmd template skeleton
}
