% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{get_treatment_colors}
\alias{get_treatment_colors}
\title{Get consistent treatment colors}
\usage{
get_treatment_colors(treatments)
}
\arguments{
\item{treatments}{Character vector of treatment names}
}
\value{
Named vector of colors
}
\description{
Get consistent treatment colors
}
