% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-process.R
\name{get_weights}
\alias{get_weights}
\title{Get all possible weights for each endpoint in graphical testing procedure}
\usage{
get_weights(graph, digits = 6, and_symbol = ", ")
}
\arguments{
\item{graph}{Graph object}

\item{digits}{Number of digits for rounding}

\item{and_symbol}{Symbol for joining}
}
\value{
Data frame of weights and scenarios
}
\description{
Get all possible weights for each endpoint in graphical testing procedure
}
