% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_MT_in_grSeq.R
\name{getDelta}
\alias{getDelta}
\title{Get a parameter under H1 on a natural scale}
\usage{
getDelta(x)
}
\arguments{
\item{x}{A endpoint object}
}
\value{
A value of parameter on a natural scale under H1
}
\description{
Get a parameter under H1 on a natural scale
}
\examples{
getDelta( x = structure(list(p1 = 0.04, p2=0.03, dropoutrate = 0.001), class="tte_exp"))
}
\keyword{internal}
