% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_MT_in_grSeq.R
\name{getEffectSizeDetails}
\alias{getEffectSizeDetails}
\title{Format a string reporting effect}
\usage{
getEffectSizeDetails(x)
}
\arguments{
\item{x}{An endpoint object}
}
\value{
A character vector
}
\description{
Format a string reporting effect
}
\examples{
getEffectSizeDetails( x = structure(list(p1 = 0.40, p2 = 0), class = "normal"))
}
\keyword{internal}
