% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_MT_in_grSeq.R
\name{getStandardizingCoef}
\alias{getStandardizingCoef}
\title{Compute standardization factor, psi,  that link natural parameter to the effect size, theta}
\usage{
getStandardizingCoef(x, ratio = 1)
}
\arguments{
\item{x}{An endpoint object}

\item{ratio}{Allocation ratio}
}
\value{
A value of standardization factor
}
\description{
Compute standardization factor, psi,  that link natural parameter to the effect size, theta
}
\examples{
getStandardizingCoef(list(list(),class="tte_exp"),ratio=1)
}
