% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_MT_in_grSeq.R
\name{getDelta}
\alias{getDelta}
\alias{getDelta.default}
\alias{getDelta.tte_exp}
\alias{getEffectSizeDetails}
\alias{getEffectSizeDetails.normal}
\alias{getEffectSizeDetails.binomial}
\alias{getEffectSizeDetails.binomial_pooled}
\alias{getEffectSizeDetails.binomial_unpooled}
\alias{getEffectSizeDetails.tte_exp}
\alias{getStandardizingCoef}
\alias{getStandardizingCoef.default}
\alias{getStandardizingCoef.binomial}
\alias{getStandardizingCoef.binomial_pooled}
\alias{getStandardizingCoef.binomial_unpooled}
\alias{n2Time}
\alias{n2Time.default}
\alias{n2Time.tte_exp}
\alias{Time2n}
\alias{Time2n.default}
\alias{Time2n.tte_exp}
\title{appendMCP internal S3}
\usage{
getDelta(x)

\method{getDelta}{default}(x)

\method{getDelta}{tte_exp}(x)

getEffectSizeDetails(x)

\method{getEffectSizeDetails}{normal}(x)

\method{getEffectSizeDetails}{binomial}(x)

\method{getEffectSizeDetails}{binomial_pooled}(x)

\method{getEffectSizeDetails}{binomial_unpooled}(x)

\method{getEffectSizeDetails}{tte_exp}(x)

getStandardizingCoef(x, ratio = 1)

\method{getStandardizingCoef}{default}(x, ratio = 1)

\method{getStandardizingCoef}{binomial}(x, ratio = 1)

\method{getStandardizingCoef}{binomial_pooled}(x, ratio = 1)

\method{getStandardizingCoef}{binomial_unpooled}(x, ratio = 1)

n2Time(x, n, enrollment, ratio)

\method{n2Time}{default}(x, n, enrollment, ratio = 1)

\method{n2Time}{tte_exp}(x, n, enrollment, ratio = 1)

Time2n(x, T, enrollment, ratio)

\method{Time2n}{default}(x, T, enrollment, ratio = 1)

\method{Time2n}{tte_exp}(x, T, enrollment, ratio = 1)
}
\arguments{
\item{x}{An endpoint object}

\item{ratio}{Allocation ratio}

\item{n}{A sample size or number of events}

\item{enrollment}{Enrollment object}

\item{T}{Calendar time points}
}
\description{
Obtain the parameter under H1 on the natural scale

Format a string reporting the effect

Compute standardization factor, psi, that links
natural parameter to the effect size, theta

Obtain a calendar time given the sample size (or events)

Calculate the sample size (or the number of events) available
 at the given calendar time
}
\examples{
\dontshow{if (FALSE) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
# Internally dispatched examples (for development or internal tests)
getStandardizingCoef(list(list(), class = "tte_exp"), ratio = 1)
# should return 0.5
\dontshow{\}) # examplesIf}
\dontshow{if (FALSE) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
\dontrun{
  Time2n(x, T, enrollment, ratio)
}
\dontshow{\}) # examplesIf}
}
\keyword{internal}
