% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/templates.R
\name{list_config_repository}
\alias{list_config_repository}
\title{Get available config repository entries}
\usage{
list_config_repository()
}
\value{
Character vector of available config names
}
\description{
Get available config repository entries
}
