% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/templates.R
\name{list_rmd_template_repository}
\alias{list_rmd_template_repository}
\title{Get available Rmd template repository entries}
\usage{
list_rmd_template_repository()
}
\value{
Character vector of available rmd_template names
}
\description{
Get available Rmd template repository entries
}
