% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/templates.R
\name{load_config}
\alias{load_config}
\title{Load study configuration from file path}
\usage{
load_config(file_path)
}
\arguments{
\item{file_path}{Full path to configuration R file}
}
\value{
Study configuration list
}
\description{
Load study configuration from file path
}
