% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/templates.R
\name{load_config_from_repository}
\alias{load_config_from_repository}
\title{Load config from repository}
\usage{
load_config_from_repository(config_name)
}
\arguments{
\item{config_name}{Name of config to load from repository}
}
\value{
Study configuration list
}
\description{
Load config from repository
}
