% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/templates.R
\name{load_rmd_template}
\alias{load_rmd_template}
\title{Load Rmd template from file path}
\usage{
load_rmd_template(file_path)
}
\arguments{
\item{file_path}{Full path to rmd_template Rmd file}
}
\value{
Path to rmd_template file (validated)
}
\description{
Load Rmd template from file path
}
