% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/templates.R
\name{load_rmd_template_from_repository}
\alias{load_rmd_template_from_repository}
\title{Load Rmd template from repository}
\usage{
load_rmd_template_from_repository(rmd_template_name)
}
\arguments{
\item{rmd_template_name}{Name of rmd_template to load from repository}
}
\value{
Path to rmd_template file
}
\description{
Load Rmd template from repository
}
