% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_MT_in_grSeq.R
\name{n2Time}
\alias{n2Time}
\title{Get calendar time given sample size (or number events)}
\usage{
n2Time(x, n, enrollment, ratio)
}
\arguments{
\item{x}{An endpoint object}

\item{n}{A sample size or number of events}

\item{enrollment}{A enrollement object}

\item{ratio}{An allocation ration}
}
\description{
Get calendar time given sample size (or number events)
}
