% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_distribution_bin}
\alias{plot_distribution_bin}
\title{Plot binary endpoint rates}
\usage{
plot_distribution_bin(distribution_bin)
}
\arguments{
\item{distribution_bin}{Binary distribution data frame with columns: endpoint, stratum, treatment, rate}
}
\value{
ggplot object
}
\description{
Plot binary endpoint rates
}
