% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_distribution_bin_rd}
\alias{plot_distribution_bin_rd}
\title{Plot risk difference for binary endpoints}
\usage{
plot_distribution_bin_rd(hypotheses)
}
\arguments{
\item{hypotheses}{Hypotheses object from processed config}
}
\value{
ggplot object
}
\description{
Plot risk difference for binary endpoints
}
