% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_distribution_tte_ahr}
\alias{plot_distribution_tte_ahr}
\title{Plot average hazard ratio over time by hypothesis}
\usage{
plot_distribution_tte_ahr(hypotheses, enroll_rate, time_grid = seq(0, 60, 0.5))
}
\arguments{
\item{hypotheses}{Hypotheses object from processed config}

\item{enroll_rate}{Enrollment rate data frame}

\item{time_grid}{Time sequence for plotting}
}
\value{
ggplot object
}
\description{
Plot average hazard ratio over time by hypothesis
}
