% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_distribution_tte_cumhaz}
\alias{plot_distribution_tte_cumhaz}
\title{Plot cumulative hazard}
\usage{
plot_distribution_tte_cumhaz(distribution_tte, time_grid = seq(0, 60, 0.5))
}
\arguments{
\item{distribution_tte}{Time-to-event distribution data frame}

\item{time_grid}{Time sequence for plotting}
}
\value{
ggplot object
}
\description{
Plot cumulative hazard
}
