% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_distribution_tte_hr}
\alias{plot_distribution_tte_hr}
\title{Plot hazard ratio over time by hypothesis}
\usage{
plot_distribution_tte_hr(hypotheses, time_grid = seq(0, 60, 0.5))
}
\arguments{
\item{hypotheses}{Hypotheses object from processed config}

\item{time_grid}{Time sequence for plotting}
}
\value{
ggplot object
}
\description{
Plot hazard ratio over time by hypothesis
}
