% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_distribution_tte_median}
\alias{plot_distribution_tte_median}
\title{Plot median survival time}
\usage{
plot_distribution_tte_median(distribution_tte)
}
\arguments{
\item{distribution_tte}{Time-to-event distribution data frame}
}
\value{
ggplot object
}
\description{
Plot median survival time
}
