% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_distribution_tte_quantiles}
\alias{plot_distribution_tte_quantiles}
\title{Plot quantiles}
\usage{
plot_distribution_tte_quantiles(distribution_tte)
}
\arguments{
\item{distribution_tte}{Time-to-event distribution data frame}
}
\value{
ggplot object
}
\description{
Plot quantiles
}
