% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_distribution_tte_weighted}
\alias{plot_distribution_tte_weighted}
\title{Plot enrollment-weighted survival by hypothesis}
\usage{
plot_distribution_tte_weighted(
  hypotheses,
  distribution_tte,
  enroll_rate,
  time_grid = seq(0, 60, 0.5),
  landmark_times = c(12, 24, 36)
)
}
\arguments{
\item{hypotheses}{Hypotheses object from processed config}

\item{distribution_tte}{Time-to-event distribution data frame}

\item{enroll_rate}{Enrollment rate data frame}

\item{time_grid}{Time sequence for plotting}

\item{landmark_times}{Times for survival rate labels}
}
\value{
ggplot object
}
\description{
Plot enrollment-weighted survival by hypothesis
}
