% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_enroll_rate}
\alias{plot_enroll_rate}
\title{Plot enrollment rates over time}
\usage{
plot_enroll_rate(enroll_rate, time_grid = NULL)
}
\arguments{
\item{enroll_rate}{Enrollment rate data frame with columns: stratum, treatments, rate, duration, ratio}

\item{time_grid}{Time sequence for plotting}
}
\value{
ggplot object
}
\description{
Plot enrollment rates over time
}
