% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_enroll_rate_cumulative}
\alias{plot_enroll_rate_cumulative}
\title{Plot cumulative enrollment over time}
\usage{
plot_enroll_rate_cumulative(enroll_rate, time_grid = NULL)
}
\arguments{
\item{enroll_rate}{Enrollment rate data frame}

\item{time_grid}{Time sequence for plotting}
}
\value{
ggplot object
}
\description{
Plot cumulative enrollment over time
}
