% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_MT_in_grSeq.R
\name{plot_iaTiming}
\alias{plot_iaTiming}
\title{Enrollment and Data Availability Plot}
\usage{
plot_iaTiming(D, enrollment, Tmax = 80, plotInPercent = TRUE)
}
\arguments{
\item{D}{A dataset containing interim analysis information}

\item{enrollment}{An enrollment object}

\item{Tmax}{A scalar defining the maximum time on the x-axis}

\item{plotInPercent}{logical; if TRUE the y-axis reports in percentages}
}
\value{
A ggplot object representing data availability for each hypothesis
}
\description{
Enrollment and Data Availability Plot
}
\examples{
\dontrun{
plot_iaTiming(D,enrollment)
}
}
\keyword{internal}
