% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_information}
\alias{plot_information}
\title{Plot information factor or fraction over time for each hypothesis}
\usage{
plot_information(config, time_grid = seq(0, 50, 1), use_fraction = FALSE)
}
\arguments{
\item{config}{Configuration list}

\item{time_grid}{Time sequence for plotting}

\item{use_fraction}{Logical, if TRUE plot information fraction, if FALSE plot information factor}
}
\value{
ggplot object
}
\description{
Plot information factor or fraction over time for each hypothesis
}
