% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_spending_functions}
\alias{plot_spending_functions}
\title{Plot alpha spending functions}
\usage{
plot_spending_functions(hypotheses, alpha = 0.025, digits = 6)
}
\arguments{
\item{hypotheses}{Object from process_config(config)$hypotheses}

\item{alpha}{Type I familywise error-rate}

\item{digits}{Number of digits for rounding}
}
\value{
ggplot object
}
\description{
Plot alpha spending functions
}
