% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process.R
\name{process_distribution}
\alias{process_distribution}
\title{Combine time-to-event and binary endpoint distributions}
\usage{
process_distribution(distribution_tte, distribution_bin)
}
\arguments{
\item{distribution_tte}{Time-to-event distribution data}

\item{distribution_bin}{Binary distribution data (can be NULL for TTE-only studies)}
}
\value{
Combined distribution data frame
}
\description{
Combine time-to-event and binary endpoint distributions
}
