% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process.R
\name{process_enroll_rate}
\alias{process_enroll_rate}
\title{Process enroll_rate}
\usage{
process_enroll_rate(enroll_rate)
}
\arguments{
\item{enroll_rate}{Configuration object}
}
\value{
Updated enroll_rate
}
\description{
Process enroll_rate
}
