% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-sim.R
\name{process_rejection_rules}
\alias{process_rejection_rules}
\title{Process rejection rules}
\usage{
process_rejection_rules(hypotheses, alpha = 0.025)
}
\arguments{
\item{hypotheses}{Data frame containing hypothesis definitions}

\item{alpha}{Overall Type I error rate}
}
\description{
Process rejection rules
}
