% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_study.R
\name{resolve_rmd_template_source}
\alias{resolve_rmd_template_source}
\title{Resolve rmd_template source (name or path)}
\usage{
resolve_rmd_template_source(rmd_template)
}
\arguments{
\item{rmd_template}{Rmd template name or file path}
}
\value{
Path to rmd_template file
}
\description{
Resolve rmd_template source (name or path)
}
