% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-sim.R
\name{sim_dataset}
\alias{sim_dataset}
\title{Dataset simulation}
\usage{
sim_dataset(
  enroll_rate,
  distribution_tte,
  distribution_bin,
  n,
  n_min_per_stratum,
  n_max_per_stratum,
  correlation_matrix,
  seed = Sys.time()
)
}
\arguments{
\item{enroll_rate}{Data frame with enrollment rates}

\item{distribution_tte}{Distribution parameters for time-to-event endpoints}

\item{distribution_bin}{Distribution parameters for binary endpoints}

\item{n}{Total sample size}

\item{n_min_per_stratum}{Minimum sample size per stratum}

\item{n_max_per_stratum}{Maximum sample size per stratum}

\item{correlation_matrix}{Correlation matrix for endpoints}

\item{seed}{Random seed for reproducibility}
}
\description{
Dataset simulation
}
