% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-sim.R
\name{sim_outcomes}
\alias{sim_outcomes}
\title{Simulation with data.table optimization}
\usage{
sim_outcomes(
  enroll_data,
  distribution_tte,
  distribution_bin,
  correlation_matrix
)
}
\arguments{
\item{enroll_data}{Data frame with enrollment data containing columns Si, Ri, Ti}

\item{distribution_tte}{Distribution parameters for time-to-event endpoints}

\item{distribution_bin}{Distribution parameters for binary endpoints}

\item{correlation_matrix}{Correlation matrix for endpoints}
}
\description{
Simulation with data.table optimization
}
