% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-tables.R
\name{table1_hux}
\alias{table1_hux}
\title{Table 1: Trial Design Summary}
\usage{
table1_hux(table1, n_digits = 0, font_size = NULL)
}
\arguments{
\item{table1}{data.frame version of table1}

\item{n_digits}{Number of decimal places for maximum events / sample size}

\item{font_size}{Font size in points, or NULL to use document default}
}
\value{
huxtable version of table1
}
\description{
Table 1: Trial Design Summary
}
