% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-tables.R
\name{table2_hux}
\alias{table2_hux}
\title{Table 2: Summary by hypothesis (with header rows)}
\usage{
table2_hux(
  table2,
  n_digits = 0,
  time_digits = 1,
  info_fraction_digits = 1,
  font_size = NULL
)
}
\arguments{
\item{table2}{data.frame version of table2}

\item{n_digits}{Number of decimal places for events / sample size}

\item{time_digits}{Number of decimal places for expected analysis time}

\item{info_fraction_digits}{Number of decimal places for information fraction}

\item{font_size}{Font size in points, or NULL to use document default}
}
\value{
huxtable version of table2
}
\description{
Table 2: Summary by hypothesis (with header rows)
}
