% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-tables.R
\name{table3_hux}
\alias{table3_hux}
\title{Table 3: Summary by criteria for conduct}
\usage{
table3_hux(
  table3,
  n_digits = 0,
  time_digits = 1,
  info_fraction_digits = 1,
  font_size = NULL
)
}
\arguments{
\item{table3}{data.frame version of table3}

\item{n_digits}{Number of decimal places for events / sample size}

\item{time_digits}{Number of decimal places for expected analysis time}

\item{info_fraction_digits}{Number of decimal places for information fraction}

\item{font_size}{Font size in points, or NULL to use document default}
}
\value{
huxtable version of table3
}
\description{
Table 3: Summary by criteria for conduct
}
