% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-tables.R
\name{table4_hux}
\alias{table4_hux}
\title{Table 4: Summary of alpha allocation by hypothesis}
\usage{
table4_hux(table4, pval_digits = 5, font_size = NULL)
}
\arguments{
\item{table4}{data.frame version of table4}

\item{pval_digits}{Number of decimal places for local alpha level}

\item{font_size}{Font size in points, or NULL to use document default}
}
\value{
huxtable version of table4
}
\description{
Table 4: Summary of alpha allocation by hypothesis
}
