% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-tables.R
\name{table5_hux}
\alias{table5_hux}
\title{Table 5: Boundary specifications}
\usage{
table5_hux(
  table5,
  info_fraction_digits = 2,
  pval_digits = 5,
  hurdle_digits = 3,
  prob_digits = 1,
  font_size = NULL
)
}
\arguments{
\item{table5}{data.frame version of table5}

\item{info_fraction_digits}{Number of decimal places for information fraction}

\item{pval_digits}{Number of decimal places for nominal p-value and local alpha level}

\item{hurdle_digits}{Number of decimal places for exit hurdle}

\item{prob_digits}{Number of decimal places for local power (as percentage)}

\item{font_size}{Font size in points, or NULL to use document default}
}
\value{
huxtable version of table5
}
\description{
Table 5: Boundary specifications
}
