% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-tables.R
\name{table6a_hux}
\alias{table6a_hux}
\title{Table 6a: Operating characteristics by analysis}
\usage{
table6a_hux(table6a, prob_digits = 1, font_size = NULL)
}
\arguments{
\item{table6a}{data.frame version of table6a}

\item{prob_digits}{Number of decimal places for probability (as percentage)}

\item{font_size}{Font size in points, or NULL to use document default}
}
\value{
huxtable version of table6a
}
\description{
Table 6a: Operating characteristics by analysis
}
