% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-tables.R
\name{table6b_hux}
\alias{table6b_hux}
\title{Table 6b: Operating characteristics across analyses}
\usage{
table6b_hux(table6b, analysis_digits = 2, time_digits = 1, font_size = NULL)
}
\arguments{
\item{table6b}{data.frame version of table6b}

\item{analysis_digits}{Number of decimal places for expected analysis index}

\item{time_digits}{Number of decimal places for expected analysis time}

\item{font_size}{Font size in points, or NULL to use document default}
}
\value{
huxtable version of table6b
}
\description{
Table 6b: Operating characteristics across analyses
}
