% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_MT_in_grSeq.R
\name{Time2n}
\alias{Time2n}
\title{Calculate sample size (or number events) available at given calendar time}
\usage{
Time2n(x, T, enrollment, ratio)
}
\arguments{
\item{x}{An endpoint object}

\item{T}{Calendar time points}

\item{enrollment}{Enrollment object}

\item{ratio}{Allocation ratio}
}
\value{
Sample size (or number events)
}
\description{
Calculate sample size (or number events) available at given calendar time
}
\examples{
\dontrun{
  Time2n(x, T, enrollment, ratio)
}
}
\keyword{internal}
