% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-sim.R
\name{update_graph}
\alias{update_graph}
\title{Update graph}
\usage{
update_graph(graph, rejected)
}
\arguments{
\item{graph}{Graph object with edges and weights}

\item{rejected}{Index of rejected hypothesis}
}
\description{
Update graph
}
