% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-sim.R
\name{update_p_thresholds}
\alias{update_p_thresholds}
\title{Update p thresholds}
\usage{
update_p_thresholds(analyses, hypotheses, graph)
}
\arguments{
\item{analyses}{Data frame containing analysis definitions}

\item{hypotheses}{Data frame containing hypothesis definitions}

\item{graph}{Graph object with edges and weights}
}
\description{
Update p thresholds
}
