% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/templates.R
\name{validate_config}
\alias{validate_config}
\title{Validate study configuration}
\usage{
validate_config(config)
}
\arguments{
\item{config}{Study configuration list}
}
\value{
TRUE if valid, otherwise throws error
}
\description{
Validate study configuration
}
