% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/appendMCP-package.R
\docType{package}
\name{appendMCP}
\alias{appendMCP-package}
\alias{appendMCP}
\title{Reporting graphical comparison procedures in GSD studies}
\description{
appendMCP provides routines for describing graphical multiple comparison
procedures (MCP) in group sequential design (GSD) studies. The objective is
to generate an R Markdown file that defines the MTP and interim analyses,
which can be appended to a statistical analysis plan document.
}
\details{
The package provides predefined templates that can be modified by a user,
ideally by changing only high-level input parameters.

To load a template in Rstudio, use "File" -> "New File" -> "R Markdown ..."
Select "From Template" (left panel), then choose "MCP in GSD"
}
\examples{
# Load template from command line:
# rmarkdown::draft("my_report", template = "mcpgen", package = "appendMCP", edit = FALSE)
# To open created file in Rstudio editor use
# rstudioapi::navigateToFile("my_report/my_report.Rmd")

}
\seealso{
Useful links:
\itemize{
  \item \url{https://ytymofyeyev.github.io/appendMCP/}
}

}
\author{
\strong{Maintainer}: Yevgen Tymofyeyev \email{ytymofye@its.jnj.com}

Authors:
\itemize{
  \item Michael Grayling \email{mgraylin@its.jnj.com}
}

}
