% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_MT_in_grSeq.R
\name{exec_calc}
\alias{exec_calc}
\title{Input Processing}
\usage{
exec_calc(inputD)
}
\arguments{
\item{inputD}{An input dataset with specifications}
}
\value{
A list with main objects to be outputted: D, ia_details and hyp_testing_dataset
}
\description{
Input Processing
}
\examples{
\dontrun{
  # Need to define inputD first
  exec_calc(inputD)
}
}
\keyword{internal}
