% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-process.R
\name{get_boundaries}
\alias{get_boundaries}
\title{Get boundary specifications for group sequential design}
\usage{
get_boundaries(
  dist_type,
  control,
  test,
  sf,
  sfpar,
  nominal,
  enroll_rate,
  distribution,
  possible_weight,
  control_pooled_rate,
  test_pooled_rate,
  information_factor,
  max_information_factor,
  information_fractions,
  times_analysed = NULL,
  test_method = "logrank",
  alpha = 0.025
)
}
\arguments{
\item{dist_type}{Type of endpoint (binary or time-to-event)}

\item{control}{Control treatment character string}

\item{test}{Test treatment character string}

\item{sf}{Spending function string (in rpact terminology)}

\item{sfpar}{Spending function parameter (if needed)}

\item{nominal}{Nominal spend information (if using)}

\item{enroll_rate}{Enrollment rate object}

\item{distribution}{Distributional information object}

\item{possible_weight}{Weight in the analysis (from the graphical test)}

\item{control_pooled_rate}{Control arm pooled rate (for binary outcomes)}

\item{test_pooled_rate}{Test arm pooled rate (for binary outcomes)}

\item{information_factor}{Information factors (either ubjects or events)}

\item{max_information_factor}{Maximum information factor (either max subjects or max events)}

\item{information_fractions}{Information fractions across the analyses}

\item{times_analysed}{Analysis times (for time-to-event endpoints)}

\item{alpha}{Type I error rate}
}
\value{
Boundary specifications object
}
\description{
Get boundary specifications for group sequential design
}
