% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_MT_in_grSeq.R
\name{knit_MT_table}
\alias{knit_MT_table}
\title{Local alpha scenarios}
\usage{
knit_MT_table(hyp_testing_dataset, digits = 5)
}
\arguments{
\item{hyp_testing_dataset}{A dataset defining testing hypotheses}

\item{digits}{A scalar defining the number of digits to report in a table}
}
\value{
A table that reports all possible scenarios for the local significance level
}
\description{
Local alpha scenarios
}
\examples{
\dontrun{
knit_MT_table(hyp_testing_dataset)
}
}
\keyword{internal}
